Initial commit
The git topic command can be used to help with a topic-branches-based
workflow. It keeps a (versioned) list of topic branches in a namespace
(default: staging/) that get merged into a "master" branch. Work in
progress patches are applied on top of that, forming the top of tree
branch called "work".

Signed-off-by: Thierry Reding <thierry.reding@gmail.com>