Add sync command
This command is used to update the local tree to the latest topic
branches from a remote repository. It will rebase the branches and base
branches, then go over each topic branch and rebase it on top of the
remote topic branch. The master and work branches will be created as
usual.

Note that this is slightly suboptimal because it doesn't detect when a
local tree has no changes on top of the remote repository and will then
create new master and work branches. They will have the same content,
but their hashes will differ.

Signed-off-by: Thierry Reding <thierry.reding@gmail.com>